restructure the code and add some minor styling
