+++ components/BookImage/style.js
@@ -1,0 +1,8 @@
+import Image from "next/image";
+import styled from "styled-components";
+
+export const StyledImage = styled(Image)`
+  max-width: 100%;
+  max-height: 100%;
+  border-radius: 4px;
+`;
