+++ components/BooksList/style.js
@@ -1,0 +1,9 @@
+import { styled } from "styled-components";
+
+export const StyledList = styled.ul`
+  display: grid;
+  gap: 20px;
+  padding-bottom: 25px;
+  padding-left: 25px;
+  list-style: none;
+`;
