+++ pages/index.js
@@ -1,9 +1,5 @@
 import BookList from "../components/BooksList/index.js";
 
 export default function HomePage() {
-  return (
-    <div>
-      <BookList />
-    </div>
-  );
+  return <BookList />;
 }
